samurai_bluebird_custos/io/image_analysis.py
# samurai_bluebird_custos/io/image_analysis.py

from PIL import ImageGrab
import numpy as np
import cv2

def capture_screenshot():
    """Capture a screenshot and return as a NumPy array."""
    screenshot = ImageGrab.grab()
    return cv2.cvtColor(np.array(screenshot), cv2.COLOR_RGB2BGR)

def analyze_screenshot_text(image):
    """Placeholder for visual analysis or scene understanding."""
    # Future: integrate ML models for object detection or sentiment analysis
    return "[Visual analysis placeholder]"
